refactor tumblr adapter into adapterbase
